adds empty list
